refactor logging setup to happen before venv creation
